Add files via upload
Quiz is an informal test or examination of a student. From our project we will try to  create a system that controls the status of the quiz. Our project is totally dependent on  the program. There are three parts of our project. First we will show question then show  some option for that question and finally we show the result of how many questions user  is answering correctly. To correct a question, we put a condition. Which is, if user puts  right answer, program will give one point otherwise zero?